requirements.txt

gradio
replicate
pillow
requests
